Refactor workflow reliability testing into modular package
