--- src/core/refactoring/workflow_reliability_testing.py
@@ -1,1162 +1,0 @@
-#!/usr/bin/env python3
-"""
-Workflow Reliability Testing System - Agent-5
-=============================================
-
-This module implements comprehensive reliability testing for automated refactoring
-workflows as part of contract REFACTOR-002 deliverables.
-
-Features:
-- Automated reliability testing
-- Stress testing and load testing
-- Failure mode analysis
-- Reliability metrics and reporting
-
-Author: Agent-5 (REFACTORING MANAGER)
-Contract: REFACTOR-002 - Automated Refactoring Workflow Implementation
-Status: IN PROGRESS
-"""
-
-import os
-import sys
-import json
-import logging
-import asyncio
-import random
-import time
-from pathlib import Path
-from typing import Dict, List, Any, Optional, Callable, Tuple
-from dataclasses import dataclass, field
-from enum import Enum
-from datetime import datetime, timedelta
-import traceback
-import statistics
-import concurrent.futures
-
-# Add src to path for imports
-sys.path.append(str(Path(__file__).parent.parent.parent))
-
-from core.managers.base_manager import BaseManager
-from core.workflow_validation import WorkflowValidationSystem, ValidationLevel
-
-# Configure logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-
-class TestType(Enum):
-    """Types of reliability tests."""
-    FUNCTIONAL = "functional"
-    STRESS = "stress"
-    LOAD = "load"
-    FAILURE_MODE = "failure_mode"
-    RECOVERY = "recovery"
-    CONSISTENCY = "consistency"
-    PERFORMANCE = "performance"
-    INTEGRATION = "integration"
-
-
-class TestResult(Enum):
-    """Test result enumeration."""
-    PASSED = "passed"
-    FAILED = "failed"
-    WARNING = "warning"
-    ERROR = "error"
-    TIMEOUT = "timeout"
-
-
-@dataclass
-class ReliabilityTest:
-    """Individual reliability test definition."""
-    test_id: str
-    name: str
-    description: str
-    test_type: TestType
-    test_func: Callable
-    parameters: Dict[str, Any] = field(default_factory=dict)
-    timeout: float = 60.0  # seconds
-    retry_count: int = 3
-    weight: float = 1.0
-    dependencies: List[str] = field(default_factory=list)
-
-
-@dataclass
-class TestExecutionResult:
-    """Result of a single test execution."""
-    test_id: str
-    test_name: str
-    test_type: TestType
-    result: TestResult
-    execution_time: float = 0.0
-    retry_count: int = 0
-    error_message: Optional[str] = None
-    performance_metrics: Dict[str, Any] = field(default_factory=dict)
-    reliability_score: float = 0.0
-    timestamp: datetime = field(default_factory=datetime.now)
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class ReliabilityTestSuite:
-    """Complete reliability test suite."""
-    suite_id: str
-    name: str
-    description: str
-    tests: List[ReliabilityTest]
-    start_time: datetime
-    end_time: Optional[datetime] = None
-    total_tests: int = 0
-    passed_tests: int = 0
-    failed_tests: int = 0
-    warning_tests: int = 0
-    error_tests: int = 0
-    timeout_tests: int = 0
-    overall_reliability: float = 0.0
-    performance_score: float = 0.0
-    stability_score: float = 0.0
-    test_results: List[TestExecutionResult] = field(default_factory=list)
-    recommendations: List[str] = field(default_factory=list)
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-
-class WorkflowReliabilityTesting(BaseManager):
-    """
-    Comprehensive workflow reliability testing system.
-    
-    This class provides automated testing, stress testing, and reliability analysis
-    for automated refactoring workflows.
-    """
-    
-    def __init__(self, config: Optional[Dict[str, Any]] = None):
-        """Initialize the workflow reliability testing system."""
-        super().__init__(config or {})
-        self.reliability_tests: Dict[str, ReliabilityTest] = {}
-        self.test_suites: Dict[str, ReliabilityTestSuite] = {}
-        self.performance_baselines: Dict[str, float] = {}
-        self.reliability_history: List[float] = []
-        self.validation_system: Optional[WorkflowValidationSystem] = None
-        
-        self._initialize_reliability_tests()
-        self._setup_logging()
-    
-    def _initialize_reliability_tests(self):
-        """Initialize comprehensive reliability tests."""
-        # Functional Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="functional_basic_workflow",
-                name="Basic Workflow Functionality",
-                description="Test basic workflow execution functionality",
-                test_type=TestType.FUNCTIONAL,
-                test_func=self._test_basic_workflow_functionality,
-                weight=2.0
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="functional_workflow_completion",
-                name="Workflow Completion",
-                description="Test complete workflow execution from start to finish",
-                test_type=TestType.FUNCTIONAL,
-                test_func=self._test_workflow_completion,
-                weight=2.5
-            )
-        )
-        
-        # Stress Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="stress_high_load",
-                name="High Load Stress Test",
-                description="Test workflow performance under high load conditions",
-                test_type=TestType.STRESS,
-                test_func=self._test_high_load_stress,
-                parameters={"concurrent_workflows": 10, "duration_minutes": 5},
-                weight=1.8
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="stress_memory_pressure",
-                name="Memory Pressure Stress Test",
-                description="Test workflow behavior under memory pressure",
-                test_type=TestType.STRESS,
-                test_func=self._test_memory_pressure,
-                parameters={"memory_limit_mb": 512},
-                weight=1.5
-            )
-        )
-        
-        # Load Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="load_concurrent_execution",
-                name="Concurrent Execution Load Test",
-                description="Test multiple workflows executing concurrently",
-                test_type=TestType.LOAD,
-                test_func=self._test_concurrent_execution,
-                parameters={"max_concurrent": 20, "total_workflows": 100},
-                weight=1.6
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="load_sequential_execution",
-                name="Sequential Execution Load Test",
-                description="Test sequential workflow execution",
-                test_type=TestType.LOAD,
-                test_func=self._test_sequential_execution,
-                parameters={"total_workflows": 50},
-                weight=1.4
-            )
-        )
-        
-        # Failure Mode Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="failure_mode_invalid_input",
-                name="Invalid Input Failure Mode Test",
-                description="Test workflow behavior with invalid inputs",
-                test_type=TestType.FAILURE_MODE,
-                test_func=self._test_invalid_input_failure,
-                weight=1.7
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="failure_mode_resource_exhaustion",
-                name="Resource Exhaustion Failure Mode Test",
-                description="Test workflow behavior under resource exhaustion",
-                test_type=TestType.FAILURE_MODE,
-                test_func=self._test_resource_exhaustion_failure,
-                weight=1.3
-            )
-        )
-        
-        # Recovery Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="recovery_workflow_restart",
-                name="Workflow Restart Recovery Test",
-                description="Test workflow recovery after failure and restart",
-                test_type=TestType.RECOVERY,
-                test_func=self._test_workflow_restart_recovery,
-                weight=1.9
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="recovery_state_persistence",
-                name="State Persistence Recovery Test",
-                description="Test workflow state persistence and recovery",
-                test_type=TestType.RECOVERY,
-                test_func=self._test_state_persistence_recovery,
-                weight=1.6
-            )
-        )
-        
-        # Consistency Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="consistency_result_reproducibility",
-                name="Result Reproducibility Consistency Test",
-                description="Test workflow result consistency across multiple runs",
-                test_type=TestType.CONSISTENCY,
-                test_func=self._test_result_reproducibility,
-                parameters={"runs": 5},
-                weight=1.8
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="consistency_behavior_uniformity",
-                name="Behavior Uniformity Consistency Test",
-                description="Test uniform behavior across different conditions",
-                test_type=TestType.CONSISTENCY,
-                test_func=self._test_behavior_uniformity,
-                weight=1.5
-            )
-        )
-        
-        # Performance Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="performance_execution_speed",
-                name="Execution Speed Performance Test",
-                description="Test workflow execution speed and performance",
-                test_type=TestType.PERFORMANCE,
-                test_func=self._test_execution_speed,
-                weight=1.4
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="performance_scalability",
-                name="Scalability Performance Test",
-                description="Test workflow scalability with increasing load",
-                test_type=TestType.PERFORMANCE,
-                test_func=self._test_scalability,
-                parameters={"scale_factors": [1, 2, 4, 8]},
-                weight=1.6
-            )
-        )
-        
-        # Integration Tests
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="integration_validation_system",
-                name="Validation System Integration Test",
-                description="Test integration with workflow validation system",
-                test_type=TestType.INTEGRATION,
-                test_func=self._test_validation_system_integration,
-                weight=1.7
-            )
-        )
-        
-        self._add_reliability_test(
-            ReliabilityTest(
-                test_id="integration_performance_benchmark",
-                name="Performance Benchmark Integration Test",
-                description="Test integration with performance benchmarking",
-                test_type=TestType.INTEGRATION,
-                test_func=self._test_performance_benchmark_integration,
-                weight=1.4
-            )
-        )
-    
-    def _add_reliability_test(self, test: ReliabilityTest):
-        """Add a reliability test to the system."""
-        self.reliability_tests[test.test_id] = test
-        logger.info(f"Added reliability test: {test.name}")
-    
-    def set_validation_system(self, validation_system: WorkflowValidationSystem):
-        """Set the workflow validation system for integration testing."""
-        self.validation_system = validation_system
-        logger.info("Workflow validation system integrated")
-    
-    async def run_reliability_test_suite(self, suite_name: str, 
-                                       test_types: Optional[List[TestType]] = None,
-                                       target_workflow_id: Optional[str] = None) -> ReliabilityTestSuite:
-        """
-        Run a comprehensive reliability test suite.
-        
-        Args:
-            suite_name: Name of the test suite
-            test_types: Optional list of test types to include
-            target_workflow_id: Optional target workflow ID for testing
-            
-        Returns:
-            Complete test suite results
-        """
-        suite_id = f"reliability_suite_{suite_name}_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
-        
-        # Filter tests based on types
-        if test_types:
-            applicable_tests = [test for test in self.reliability_tests.values() 
-                              if test.test_type in test_types]
-        else:
-            applicable_tests = list(self.reliability_tests.values())
-        
-        # Create test suite
-        test_suite = ReliabilityTestSuite(
-            suite_id=suite_id,
-            name=suite_name,
-            description=f"Reliability test suite for {suite_name}",
-            tests=applicable_tests,
-            start_time=datetime.now()
-        )
-        
-        test_suite.total_tests = len(applicable_tests)
-        logger.info(f"Starting reliability test suite: {suite_name} with {len(applicable_tests)} tests")
-        
-        # Execute tests
-        for test in applicable_tests:
-            try:
-                result = await self._execute_reliability_test(test, target_workflow_id)
-                test_suite.test_results.append(result)
-                
-                # Update counters
-                if result.result == TestResult.PASSED:
-                    test_suite.passed_tests += 1
-                elif result.result == TestResult.FAILED:
-                    test_suite.failed_tests += 1
-                elif result.result == TestResult.WARNING:
-                    test_suite.warning_tests += 1
-                elif result.result == TestResult.ERROR:
-                    test_suite.error_tests += 1
-                elif result.result == TestResult.TIMEOUT:
-                    test_suite.timeout_tests += 1
-                    
-            except Exception as e:
-                logger.error(f"Reliability test {test.test_id} failed with error: {str(e)}")
-                error_result = TestExecutionResult(
-                    test_id=test.test_id,
-                    test_name=test.name,
-                    test_type=test.test_type,
-                    result=TestResult.ERROR,
-                    error_message=str(e),
-                    timestamp=datetime.now()
-                )
-                test_suite.test_results.append(error_result)
-                test_suite.error_tests += 1
-        
-        # Calculate scores
-        test_suite = self._calculate_test_suite_scores(test_suite)
-        
-        # Generate recommendations
-        test_suite.recommendations = self._generate_test_recommendations(test_suite)
-        
-        # Finalize suite
-        test_suite.end_time = datetime.now()
-        
-        # Store test suite
-        self.test_suites[suite_id] = test_suite
-        
-        # Update reliability history
-        self.reliability_history.append(test_suite.overall_reliability)
-        
-        logger.info(f"Reliability test suite {suite_name} completed")
-        logger.info(f"Overall reliability: {test_suite.overall_reliability:.2f}%")
-        
-        return test_suite
-    
-    async def _execute_reliability_test(self, test: ReliabilityTest, 
-                                      target_workflow_id: Optional[str] = None) -> TestExecutionResult:
-        """Execute a single reliability test."""
-        start_time = time.time()
-        retry_count = 0
-        last_error = None
-        
-        # Execute test with retries
-        while retry_count <= test.retry_count:
-            try:
-                if asyncio.iscoroutinefunction(test.test_func):
-                    result = await asyncio.wait_for(
-                        test.test_func(target_workflow_id, test.parameters),
-                        timeout=test.timeout
-                    )
-                else:
-                    result = test.test_func(target_workflow_id, test.parameters)
-                
-                execution_time = time.time() - start_time
-                
-                # Create test result
-                test_result = TestExecutionResult(
-                    test_id=test.test_id,
-                    test_name=test.name,
-                    test_type=test.test_type,
-                    result=result.get("status", TestResult.PASSED),
-                    execution_time=execution_time,
-                    retry_count=retry_count,
-                    performance_metrics=result.get("performance_metrics", {}),
-                    reliability_score=result.get("reliability_score", 100.0),
-                    timestamp=datetime.now(),
-                    metadata=result.get("metadata", {})
-                )
-                
-                return test_result
-                
-            except asyncio.TimeoutError:
-                last_error = "Test execution timed out"
-                retry_count += 1
-                if retry_count <= test.retry_count:
-                    await asyncio.sleep(1)  # Brief pause before retry
-                    continue
-                else:
-                    execution_time = time.time() - start_time
-                    return TestExecutionResult(
-                        test_id=test.test_id,
-                        test_name=test.name,
-                        test_type=test.test_type,
-                        result=TestResult.TIMEOUT,
-                        execution_time=execution_time,
-                        retry_count=retry_count,
-                        error_message=last_error,
-                        timestamp=datetime.now()
-                    )
-                    
-            except Exception as e:
-                last_error = str(e)
-                retry_count += 1
-                if retry_count <= test.retry_count:
-                    await asyncio.sleep(1)  # Brief pause before retry
-                    continue
-                else:
-                    execution_time = time.time() - start_time
-                    return TestExecutionResult(
-                        test_id=test.test_id,
-                        test_name=test.name,
-                        test_type=test.test_type,
-                        result=TestResult.ERROR,
-                        execution_time=execution_time,
-                        retry_count=retry_count,
-                        error_message=last_error,
-                        timestamp=datetime.now()
-                    )
-    
-    def _calculate_test_suite_scores(self, test_suite: ReliabilityTestSuite) -> ReliabilityTestSuite:
-        """Calculate comprehensive test suite scores."""
-        if not test_suite.test_results:
-            return test_suite
-        
-        # Calculate overall reliability (weighted average)
-        total_weight = 0.0
-        weighted_reliability = 0.0
-        
-        for result in test_suite.test_results:
-            test = self.reliability_tests.get(result.test_id)
-            if test:
-                weight = test.weight
-                total_weight += weight
-                weighted_reliability += result.reliability_score * weight
-        
-        if total_weight > 0:
-            test_suite.overall_reliability = weighted_reliability / total_weight
-        
-        # Calculate performance score (based on execution times)
-        execution_times = [r.execution_time for r in test_suite.test_results if r.execution_time > 0]
-        if execution_times:
-            avg_time = statistics.mean(execution_times)
-            # Normalize to 0-100 scale (faster = higher score)
-            test_suite.performance_score = max(0, 100 - (avg_time * 2))
-        
-        # Calculate stability score (based on consistency of results)
-        if test_suite.total_tests > 0:
-            stability_factors = {
-                TestResult.PASSED: 1.0,
-                TestResult.WARNING: 0.8,
-                TestResult.FAILED: 0.3,
-                TestResult.ERROR: 0.0,
-                TestResult.TIMEOUT: 0.2
-            }
-            
-            stability_score = 0.0
-            for result in test_suite.test_results:
-                factor = stability_factors.get(result.result, 0.0)
-                stability_score += factor
-            
-            test_suite.stability_score = (stability_score / test_suite.total_tests) * 100
-        
-        return test_suite
-    
-    def _generate_test_recommendations(self, test_suite: ReliabilityTestSuite) -> List[str]:
-        """Generate actionable recommendations based on test results."""
-        recommendations = []
-        
-        # Analyze failed tests
-        failed_tests = [r for r in test_suite.test_results if r.result == TestResult.FAILED]
-        if failed_tests:
-            recommendations.append(f"Address {len(failed_tests)} failed reliability tests to improve overall reliability")
-            
-            for result in failed_tests[:3]:  # Top 3 failures
-                recommendations.append(f"Fix {result.test_name}: {result.error_message or 'Review implementation'}")
-        
-        # Analyze error tests
-        error_tests = [r for r in test_suite.test_results if r.result == TestResult.ERROR]
-        if error_tests:
-            recommendations.append(f"Investigate {len(error_tests)} error conditions for system stability improvements")
-        
-        # Performance recommendations
-        if test_suite.performance_score < 70:
-            recommendations.append("Optimize test execution times for better performance scores")
-        
-        # Stability recommendations
-        if test_suite.stability_score < 80:
-            recommendations.append("Improve test consistency and error handling for better stability")
-        
-        # Reliability recommendations
-        if test_suite.overall_reliability < 85:
-            recommendations.append("Enhance test implementations and error recovery for higher reliability")
-        
-        # General improvement recommendations
-        if test_suite.overall_reliability < 90:
-            recommendations.append("Consider implementing additional reliability tests for comprehensive coverage")
-        
-        if not recommendations:
-            recommendations.append("Excellent reliability test results! Maintain current quality standards.")
-        
-        return recommendations
-    
-    # Reliability Test Implementations
-    
-    async def _test_basic_workflow_functionality(self, target_workflow_id: Optional[str], 
-                                               parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test basic workflow execution functionality."""
-        await asyncio.sleep(2.0)  # Simulate test execution
-        
-        # Mock test logic
-        functionality_score = 95.0
-        tests_passed = 8
-        total_tests = 10
-        
-        status = TestResult.PASSED if functionality_score >= 90 else TestResult.FAILED
-        
-        return {
-            "status": status,
-            "reliability_score": functionality_score,
-            "performance_metrics": {
-                "tests_passed": tests_passed,
-                "total_tests": total_tests,
-                "success_rate": f"{tests_passed/total_tests*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "basic_functionality",
-                "complexity": "low"
-            }
-        }
-    
-    async def _test_workflow_completion(self, target_workflow_id: Optional[str], 
-                                       parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test complete workflow execution from start to finish."""
-        await asyncio.sleep(3.5)  # Simulate longer test execution
-        
-        # Mock test logic
-        completion_score = 92.0
-        workflow_stages = 5
-        completed_stages = 5
-        
-        status = TestResult.PASSED if completion_score >= 90 else TestResult.FAILED
-        
-        return {
-            "status": status,
-            "reliability_score": completion_score,
-            "performance_metrics": {
-                "workflow_stages": workflow_stages,
-                "completed_stages": completed_stages,
-                "completion_rate": "100%"
-            },
-            "metadata": {
-                "test_category": "end_to_end",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_high_load_stress(self, target_workflow_id: Optional[str], 
-                                    parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow performance under high load conditions."""
-        concurrent_workflows = parameters.get("concurrent_workflows", 10)
-        duration_minutes = parameters.get("duration_minutes", 5)
-        
-        await asyncio.sleep(duration_minutes * 0.1)  # Simulate proportional test time
-        
-        # Mock stress test logic
-        stress_score = 88.0
-        throughput = 150  # workflows per minute
-        error_rate = 2.5  # percentage
-        
-        status = TestResult.PASSED if stress_score >= 80 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": stress_score,
-            "performance_metrics": {
-                "concurrent_workflows": concurrent_workflows,
-                "throughput": f"{throughput} workflows/min",
-                "error_rate": f"{error_rate}%",
-                "duration_minutes": duration_minutes
-            },
-            "metadata": {
-                "test_category": "stress_testing",
-                "complexity": "high"
-            }
-        }
-    
-    async def _test_memory_pressure(self, target_workflow_id: Optional[str], 
-                                   parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow behavior under memory pressure."""
-        memory_limit_mb = parameters.get("memory_limit_mb", 512)
-        
-        await asyncio.sleep(2.5)  # Simulate memory pressure test
-        
-        # Mock memory test logic
-        memory_score = 85.0
-        peak_memory_mb = 480
-        memory_efficiency = "good"
-        
-        status = TestResult.PASSED if memory_score >= 80 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": memory_score,
-            "performance_metrics": {
-                "memory_limit_mb": memory_limit_mb,
-                "peak_memory_mb": peak_memory_mb,
-                "memory_efficiency": memory_efficiency,
-                "utilization": f"{peak_memory_mb/memory_limit_mb*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "resource_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_concurrent_execution(self, target_workflow_id: Optional[str], 
-                                        parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test multiple workflows executing concurrently."""
-        max_concurrent = parameters.get("max_concurrent", 20)
-        total_workflows = parameters.get("total_workflows", 100)
-        
-        await asyncio.sleep(4.0)  # Simulate concurrent execution test
-        
-        # Mock concurrent test logic
-        concurrent_score = 90.0
-        actual_concurrent = 18
-        total_completed = 98
-        
-        status = TestResult.PASSED if concurrent_score >= 85 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": concurrent_score,
-            "performance_metrics": {
-                "max_concurrent": max_concurrent,
-                "actual_concurrent": actual_concurrent,
-                "total_workflows": total_workflows,
-                "completed_workflows": total_completed,
-                "success_rate": f"{total_completed/total_workflows*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "concurrency_testing",
-                "complexity": "high"
-            }
-        }
-    
-    async def _test_sequential_execution(self, target_workflow_id: Optional[str], 
-                                       parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test sequential workflow execution."""
-        total_workflows = parameters.get("total_workflows", 50)
-        
-        await asyncio.sleep(3.0)  # Simulate sequential execution test
-        
-        # Mock sequential test logic
-        sequential_score = 94.0
-        completed_workflows = 50
-        average_time = 2.5  # seconds per workflow
-        
-        status = TestResult.PASSED if sequential_score >= 90 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": sequential_score,
-            "performance_metrics": {
-                "total_workflows": total_workflows,
-                "completed_workflows": completed_workflows,
-                "success_rate": "100%",
-                "average_time_per_workflow": f"{average_time}s"
-            },
-            "metadata": {
-                "test_category": "sequential_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_invalid_input_failure(self, target_workflow_id: Optional[str], 
-                                        parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow behavior with invalid inputs."""
-        await asyncio.sleep(2.0)  # Simulate invalid input test
-        
-        # Mock failure mode test logic
-        failure_score = 87.0
-        invalid_inputs_tested = 15
-        graceful_failures = 14
-        
-        status = TestResult.PASSED if failure_score >= 80 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": failure_score,
-            "performance_metrics": {
-                "invalid_inputs_tested": invalid_inputs_tested,
-                "graceful_failures": graceful_failures,
-                "failure_handling_rate": f"{graceful_failures/invalid_inputs_tested*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "failure_mode_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_resource_exhaustion_failure(self, target_workflow_id: Optional[str], 
-                                              parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow behavior under resource exhaustion."""
-        await asyncio.sleep(2.5)  # Simulate resource exhaustion test
-        
-        # Mock resource exhaustion test logic
-        exhaustion_score = 82.0
-        resource_scenarios = 8
-        graceful_degradation = 7
-        
-        status = TestResult.PASSED if exhaustion_score >= 80 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": exhaustion_score,
-            "performance_metrics": {
-                "resource_scenarios": resource_scenarios,
-                "graceful_degradation": graceful_degradation,
-                "degradation_rate": f"{graceful_degradation/resource_scenarios*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "failure_mode_testing",
-                "complexity": "high"
-            }
-        }
-    
-    async def _test_workflow_restart_recovery(self, target_workflow_id: Optional[str], 
-                                            parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow recovery after failure and restart."""
-        await asyncio.sleep(3.0)  # Simulate restart recovery test
-        
-        # Mock recovery test logic
-        recovery_score = 89.0
-        restart_scenarios = 10
-        successful_recoveries = 9
-        
-        status = TestResult.PASSED if recovery_score >= 85 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": recovery_score,
-            "performance_metrics": {
-                "restart_scenarios": restart_scenarios,
-                "successful_recoveries": successful_recoveries,
-                "recovery_rate": f"{successful_recoveries/restart_scenarios*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "recovery_testing",
-                "complexity": "high"
-            }
-        }
-    
-    async def _test_state_persistence_recovery(self, target_workflow_id: Optional[str], 
-                                             parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow state persistence and recovery."""
-        await asyncio.sleep(2.5)  # Simulate state persistence test
-        
-        # Mock state persistence test logic
-        persistence_score = 91.0
-        state_checkpoints = 12
-        successful_persistence = 12
-        
-        status = TestResult.PASSED if persistence_score >= 90 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": persistence_score,
-            "performance_metrics": {
-                "state_checkpoints": state_checkpoints,
-                "successful_persistence": successful_persistence,
-                "persistence_rate": "100%"
-            },
-            "metadata": {
-                "test_category": "recovery_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_result_reproducibility(self, target_workflow_id: Optional[str], 
-                                         parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow result consistency across multiple runs."""
-        runs = parameters.get("runs", 5)
-        
-        await asyncio.sleep(runs * 0.8)  # Simulate multiple runs
-        
-        # Mock reproducibility test logic
-        reproducibility_score = 93.0
-        consistent_results = 4
-        total_runs = runs
-        
-        status = TestResult.PASSED if reproducibility_score >= 90 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": reproducibility_score,
-            "performance_metrics": {
-                "total_runs": total_runs,
-                "consistent_results": consistent_results,
-                "consistency_rate": f"{consistent_results/total_runs*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "consistency_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_behavior_uniformity(self, target_workflow_id: Optional[str], 
-                                      parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test uniform behavior across different conditions."""
-        await asyncio.sleep(2.5)  # Simulate behavior uniformity test
-        
-        # Mock uniformity test logic
-        uniformity_score = 88.0
-        test_conditions = 15
-        uniform_behavior = 13
-        
-        status = TestResult.PASSED if uniformity_score >= 85 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": uniformity_score,
-            "performance_metrics": {
-                "test_conditions": test_conditions,
-                "uniform_behavior": uniform_behavior,
-                "uniformity_rate": f"{uniform_behavior/test_conditions*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "consistency_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_execution_speed(self, target_workflow_id: Optional[str], 
-                                  parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow execution speed and performance."""
-        await asyncio.sleep(2.0)  # Simulate speed test
-        
-        # Mock speed test logic
-        speed_score = 92.0
-        baseline_time = 10.0  # seconds
-        optimized_time = 7.5  # seconds
-        improvement = 25.0  # percentage
-        
-        status = TestResult.PASSED if speed_score >= 90 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": speed_score,
-            "performance_metrics": {
-                "baseline_time": f"{baseline_time}s",
-                "optimized_time": f"{optimized_time}s",
-                "improvement": f"{improvement}%",
-                "speedup_factor": f"{baseline_time/optimized_time:.2f}x"
-            },
-            "metadata": {
-                "test_category": "performance_testing",
-                "complexity": "low"
-            }
-        }
-    
-    async def _test_scalability(self, target_workflow_id: Optional[str], 
-                               parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test workflow scalability with increasing load."""
-        scale_factors = parameters.get("scale_factors", [1, 2, 4, 8])
-        
-        await asyncio.sleep(len(scale_factors) * 1.5)  # Simulate scalability test
-        
-        # Mock scalability test logic
-        scalability_score = 89.0
-        linear_scaling = 3
-        total_factors = len(scale_factors)
-        
-        status = TestResult.PASSED if scalability_score >= 85 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": scalability_score,
-            "performance_metrics": {
-                "scale_factors": scale_factors,
-                "linear_scaling": linear_scaling,
-                "total_factors": total_factors,
-                "scaling_efficiency": f"{linear_scaling/total_factors*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "performance_testing",
-                "complexity": "high"
-            }
-        }
-    
-    async def _test_validation_system_integration(self, target_workflow_id: Optional[str], 
-                                                parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test integration with workflow validation system."""
-        await asyncio.sleep(2.5)  # Simulate integration test
-        
-        # Mock integration test logic
-        integration_score = 91.0
-        validation_rules = 12
-        successful_validations = 11
-        
-        status = TestResult.PASSED if integration_score >= 90 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": integration_score,
-            "performance_metrics": {
-                "validation_rules": validation_rules,
-                "successful_validations": successful_validations,
-                "validation_success_rate": f"{successful_validations/validation_rules*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "integration_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    async def _test_performance_benchmark_integration(self, target_workflow_id: Optional[str], 
-                                                    parameters: Dict[str, Any]) -> Dict[str, Any]:
-        """Test integration with performance benchmarking."""
-        await asyncio.sleep(2.0)  # Simulate benchmark integration test
-        
-        # Mock benchmark integration test logic
-        benchmark_score = 87.0
-        benchmark_metrics = 8
-        successful_benchmarks = 7
-        
-        status = TestResult.PASSED if benchmark_score >= 85 else TestResult.WARNING
-        
-        return {
-            "status": status,
-            "reliability_score": benchmark_score,
-            "performance_metrics": {
-                "benchmark_metrics": benchmark_metrics,
-                "successful_benchmarks": successful_benchmarks,
-                "benchmark_success_rate": f"{successful_benchmarks/benchmark_metrics*100:.1f}%"
-            },
-            "metadata": {
-                "test_category": "integration_testing",
-                "complexity": "medium"
-            }
-        }
-    
-    def get_test_suite(self, suite_id: str) -> Optional[ReliabilityTestSuite]:
-        """Get a test suite by ID."""
-        return self.test_suites.get(suite_id)
-    
-    def list_test_suites(self) -> List[Dict[str, Any]]:
-        """List all test suites with summary information."""
-        suite_list = []
-        
-        for suite_id, suite in self.test_suites.items():
-            suite_info = {
-                "suite_id": suite_id,
-                "name": suite.name,
-                "overall_reliability": suite.overall_reliability,
-                "performance_score": suite.performance_score,
-                "stability_score": suite.stability_score,
-                "total_tests": suite.total_tests,
-                "passed_tests": suite.passed_tests,
-                "failed_tests": suite.failed_tests,
-                "start_time": suite.start_time.isoformat(),
-                "end_time": suite.end_time.isoformat() if suite.end_time else None
-            }
-            suite_list.append(suite_info)
-        
-        return suite_list
-    
-    def get_reliability_trends(self) -> Dict[str, Any]:
-        """Get reliability trends and statistics."""
-        if not self.reliability_history:
-            return {"message": "No reliability data available"}
-        
-        return {
-            "current_reliability": self.reliability_history[-1] if self.reliability_history else 0,
-            "average_reliability": statistics.mean(self.reliability_history),
-            "reliability_trend": "improving" if len(self.reliability_history) >= 2 and 
-                                self.reliability_history[-1] > self.reliability_history[-2] else "stable",
-            "total_test_suites": len(self.reliability_history),
-            "reliability_history": self.reliability_history[-10:]  # Last 10 values
-        }
-    
-    def export_test_suite_report(self, suite_id: str, output_path: str) -> bool:
-        """Export a test suite report to JSON."""
-        suite = self.get_test_suite(suite_id)
-        if not suite:
-            return False
-        
-        try:
-            # Prepare export data
-            export_data = {
-                "suite_id": suite.suite_id,
-                "name": suite.name,
-                "description": suite.description,
-                "start_time": suite.start_time.isoformat(),
-                "end_time": suite.end_time.isoformat() if suite.end_time else None,
-                "scores": {
-                    "overall_reliability": suite.overall_reliability,
-                    "performance_score": suite.performance_score,
-                    "stability_score": suite.stability_score
-                },
-                "test_results": [
-                    {
-                        "test_id": result.test_id,
-                        "test_name": result.test_name,
-                        "test_type": result.test_type.value,
-                        "result": result.result.value,
-                        "execution_time": result.execution_time,
-                        "retry_count": result.retry_count,
-                        "error_message": result.error_message,
-                        "reliability_score": result.reliability_score,
-                        "performance_metrics": result.performance_metrics,
-                        "timestamp": result.timestamp.isoformat()
-                    }
-                    for result in suite.test_results
-                ],
-                "recommendations": suite.recommendations,
-                "metadata": suite.metadata
-            }
-            
-            with open(output_path, 'w') as f:
-                json.dump(export_data, f, indent=2, default=str)
-            
-            logger.info(f"Test suite report exported to: {output_path}")
-            return True
-            
-        except Exception as e:
-            logger.error(f"Failed to export test suite report: {str(e)}")
-            return False
-
-
-# Example usage and testing
-async def demo_reliability_testing():
-    """Demonstrate the workflow reliability testing system."""
-    print("🧪 Workflow Reliability Testing Demo")
-    print("=" * 50)
-    
-    # Initialize reliability testing system
-    reliability_testing = WorkflowReliabilityTesting()
-    
-    # Run comprehensive reliability test suite
-    suite_name = "Comprehensive Reliability Test Suite"
-    
-    print(f"✅ Starting reliability test suite: {suite_name}")
-    
-    # Run tests for all types
-    test_suite = await reliability_testing.run_reliability_test_suite(
-        suite_name=suite_name,
-        test_types=None  # All test types
-    )
-    
-    print(f"\n📊 Test Suite Results:")
-    print(f"  Overall Reliability: {test_suite.overall_reliability:.2f}%")
-    print(f"  Performance Score: {test_suite.performance_score:.2f}%")
-    print(f"  Stability Score: {test_suite.stability_score:.2f}%")
-    print(f"  Tests Passed: {test_suite.passed_tests}/{test_suite.total_tests}")
-    
-    print(f"\n💡 Recommendations:")
-    for i, recommendation in enumerate(test_suite.recommendations[:3], 1):
-        print(f"  {i}. {recommendation}")
-    
-    # Export test suite report
-    report_path = "reliability_test_suite_report.json"
-    if reliability_testing.export_test_suite_report(test_suite.suite_id, report_path):
-        print(f"\n📊 Test suite report exported to: {report_path}")
-    
-    # Get reliability trends
-    trends = reliability_testing.get_reliability_trends()
-    print(f"\n📈 Reliability Trends:")
-    print(f"  Current: {trends['current_reliability']:.2f}%")
-    print(f"  Average: {trends['average_reliability']:.2f}%")
-    print(f"  Trend: {trends['reliability_trend']}")
-    
-    print("\n🎉 Reliability testing demo completed successfully!")
-
-
-if __name__ == "__main__":
-    # Run the demo
-    asyncio.run(demo_reliability_testing())
